move code to apps.ts
